Move QRRunnables from subclasses of PregameActivity and Climb fragment to be in its own class Scouter name should also now output correctly when viewing a cached QR code
