+++ app/src/main/java/com/mercury1089/scoutingapp2025/qr/QRUtils.java
@@ -1,0 +1,45 @@
+package com.mercury1089.scoutingapp2025.qr;
+
+
+import android.content.Context;
+import android.graphics.Bitmap;
+
+import com.google.zxing.BarcodeFormat;
+import com.google.zxing.MultiFormatWriter;
+import com.google.zxing.WriterException;
+import com.google.zxing.common.BitMatrix;
+import com.mercury1089.scoutingapp2025.PregameActivity;
+import com.mercury1089.scoutingapp2025.R;
+import com.mercury1089.scoutingapp2025.utils.GenUtils;
+
+public class QRUtils {
+    //for QR code generator
+    public final static int QRCodeSize = 500;
+    public static Bitmap textToImageEncode(Context ctx, String qrString) throws WriterException {
+        BitMatrix bitMatrix;
+        try {
+            bitMatrix = new MultiFormatWriter().encode(
+                    qrString,
+                    BarcodeFormat.QR_CODE,
+                    QRCodeSize, QRCodeSize, null
+            );
+        } catch (IllegalArgumentException illegalArgumentException) {
+            return null;
+        }
+
+        int bitMatrixWidth = bitMatrix.getWidth();
+        int bitMatrixHeight = bitMatrix.getHeight();
+        int[] pixels = new int[bitMatrixWidth * bitMatrixHeight];
+        for (int y = 0; y < bitMatrixHeight; y++) {
+            int offset = y * bitMatrixWidth;
+            for (int x = 0; x < bitMatrixWidth; x++) {
+                pixels[offset + x] = bitMatrix.get(x, y) ?
+                        GenUtils.getAColor(ctx, R.color.black) : GenUtils.getAColor(ctx, R.color.white);
+            }
+        }
+
+        Bitmap bitmap = Bitmap.createBitmap(bitMatrixWidth, bitMatrixHeight, Bitmap.Config.ARGB_4444);
+        bitmap.setPixels(pixels, 0, 500, 0, 0, bitMatrixWidth, bitMatrixHeight);
+        return bitmap;
+    }
+}
